Refactor price policy views and extract Usage concern
