+++ app/models/instrument_price_policy.rb
@@ -1,6 +1,7 @@
 class InstrumentPricePolicy < PricePolicy
 
   include InstrumentPricePolicyCalculations
+  include PricePolicies::Usage
 
   CHARGE_FOR = {
     usage: "usage",
@@ -8,9 +9,10 @@ class InstrumentPricePolicy < PricePolicy
     reservation: "reservation",
   }.freeze
 
-  validates :usage_rate, :minimum_cost, :usage_subsidy, :cancellation_cost, numericality: { allow_nil: true, greater_than_or_equal_to: 0 }
-  validates :usage_rate, presence: true, unless: :restrict_purchase?
+  validates :cancellation_cost, numericality: { allow_blank: true, greater_than_or_equal_to: 0 }
   validates :charge_for, inclusion: CHARGE_FOR.values
+
+  # Deprecated attributes used by OldPricePolicy
   validates :reservation_rate,
             :reservation_subsidy,
             :overage_rate,
@@ -27,59 +29,7 @@ class InstrumentPricePolicy < PricePolicy
     pgp.try(:reservation_window) || 0
   end
 
-  def has_rate?
-    usage_rate && usage_rate > -1
-  end
-
-  def has_minimum_cost?
-    minimum_cost && minimum_cost > -1
-  end
-
-  def free?
-    usage_rate.to_f == 0
-  end
-
-  def usage_rate=(hourly_rate)
-    super
-    self[:usage_rate] /= 60.0 if self[:usage_rate].respond_to? :/
-  end
-
-  def usage_subsidy=(hourly_subsidy)
-    super
-    self[:usage_subsidy] /= 60.0 if self[:usage_subsidy].respond_to? :/
-  end
-
-  def hourly_usage_rate
-    usage_rate.try :*, 60
-  end
-
-  def hourly_usage_subsidy
-    usage_subsidy.try :*, 60
-  end
-
-  def subsidized_hourly_usage_cost
-    hourly_usage_rate - hourly_usage_subsidy
-  end
-
-  def minimum_cost_subsidy
-    return unless has_minimum_cost?
-    minimum_cost * subsidy_ratio
-  end
-
-  def subsidized_minimum_cost
-    return unless has_minimum_cost?
-    minimum_cost - minimum_cost_subsidy
-  end
-
   private
-
-  def rate_field
-    :usage_rate
-  end
-
-  def subsidy_field
-    :usage_subsidy
-  end
 
   def ensure_reservation_window
     pgp = PriceGroupProduct.find_by_price_group_id_and_product_id(price_group.id, product.id)
