+++ app/models/price_policy.rb
@@ -9,10 +9,6 @@ class PricePolicy < ActiveRecord::Base
   validates_presence_of :start_date, :price_group_id, :type
   validate :start_date_is_unique, if: :start_date?
 
-  validates :unit_cost, :unit_subsidy, :usage_rate, :usage_subsidy,
-    :reservation_rate, :overage_rate, :overage_subsidy, :minimum_cost,
-    numericality: { greater_than_or_equal_to: 0 }, allow_nil: true
-
   validate :subsidy_less_than_rate, unless: :restrict_purchase?
 
   validates_each :expire_date do |record, _attr, value|
