+++ app/views/instrument_price_policies/_amount_row.html.haml
@@ -1,0 +1,22 @@
+%td
+  = t("price_policies.rate")
+  %br
+  = text_field_tag "price_policy_#{price_group.id}[usage_rate]",
+    display_usage_rate(price_group, price_policy), size: 8, class: "#{price_group.master_internal? ? 'master_usage_cost' : ''} usage_rate"
+
+%td
+  = t("price_policies.amount")
+  %br
+  = text_field_tag "price_policy_#{price_group.id}[minimum_cost]",
+    params["price_policy_#{price_group.id}"] ? params["price_policy_#{price_group.id}"][:minimum_cost] : number_to_currency(price_policy.minimum_cost, unit: "", delimiter: ""),
+    size: 8,
+    class: price_group.master_internal? ? "master_minimum_cost" : ""
+
+- if local_assigns[:cancellation] != false
+  %td
+    = t("price_policies.amount")
+    %br
+    = text_field_tag "price_policy_#{price_group.id}[cancellation_cost]",
+      params["price_policy_#{price_group.id}"] ? params["price_policy_#{price_group.id}"][:cancellation_cost] : number_to_currency(price_policy.cancellation_cost, unit: "", delimiter: ""),
+      size: 8,
+      class: price_group.master_internal? ? "master_cancellation_cost" : ""
