+++ app/views/instrument_price_policies/_price_policy_fields.html.haml
@@ -11,12 +11,12 @@
 %table.table.table-striped.table-hover.price-policy-table
   %thead
     %tr
-      %th= t(".price_group")
-      %th= t(".type")
-      %th= t(".can_purchase")
-      %th= t(".rate_per_hour")
-      %th= t(".minimum_cost")
-      %th= t(".reservation_cost")
+      %th= PriceGroup.model_name.human
+      %th= price_policy.class.human_attribute_name(:type)
+      %th= price_policy.class.human_attribute_name(:can_purchase)
+      %th= price_policy.class.human_attribute_name(:hourly_usage_rate)
+      %th= price_policy.class.human_attribute_name(:minimum_cost)
+      %th= price_policy.class.human_attribute_name(:cancellation_cost)
   %tbody
     - @price_policies.each do |price_policy|
       - price_group = price_policy.price_group
@@ -25,13 +25,7 @@
         %td= price_group.type_string
         %td= check_box_tag "price_policy_#{price_group.id}[can_purchase]",
           "1", price_policy.can_purchase?, class: "can_purchase"
-        - case
-        - when price_group.external?
-          = render "price_policies/instrument/external_row",
-            pg: price_group, pp: price_policy
-        - when price_group.master_internal?
-          = render "price_policies/instrument/master_internal_row",
-            pg: price_group, pp: price_policy
+        - if price_group.external? || price_group.master_internal?
+          = render "amount_row", price_group: price_group, price_policy: price_policy
         - else
-          = render "price_policies/instrument/internal_row",
-            pg: price_group, pp: price_policy
+          = render "adjustment_row", price_group: price_group, price_policy: price_policy
