+++ app/views/price_policies/index.html.haml
@@ -22,7 +22,7 @@
   %h4= text("dates", start: human_date(@current_start_date), expires: human_date(policy.expire_date))
   - if policy.charge_for.present?
     %p.charge-for= text("charges_for", charge_for: policy.charge_for)
-  = render "price_policies/#{@product.is_a?(Instrument) ? 'instrument/' : ''}table",
+  = render "table",
     price_policies: @current_price_policies,
     url_date: @current_start_date.strftime("%Y-%m-%d")
 
@@ -35,7 +35,7 @@
     %h4= text("dates", start: human_date(date), expires: human_date(policy.expire_date))
     - if policy.charge_for.present?
       %p.charge-for= text("charges_for", charge_for: policy.charge_for)
-    = render "price_policies/#{@product.is_a?(Instrument) ? 'instrument/' : ''}table",
+    = render "table",
       price_policies: rules,
       url_date: date.strftime("%Y-%m-%d")
 
@@ -48,6 +48,6 @@
     - @past_price_policies_by_date.each do |date, rules|
       - policy = rules.first
       %h4= text("dates", start: human_date(date), expires: human_date(policy.expire_date))
-      = render "price_policies/#{@product.is_a?(Instrument) ? 'instrument/' : ''}table",
+      = render "table",
         price_policies: rules,
         url_date: date.strftime("%Y-%m-%d")
