--- app/views/price_policies/instrument/_external_row.html.haml
@@ -1,19 +1,0 @@
-%td
-  = t("price_policies.rate")
-  %br
-  = text_field_tag "price_policy_#{pg.id}[usage_rate]",
-    display_usage_rate(pg, pp), size: 8, class: "usage_rate"
-
-%td
-  = t("price_policies.amount")
-  %br
-  = text_field_tag "price_policy_#{pg.id}[minimum_cost]",
-      params["price_policy_#{pg.id}"] ? params["price_policy_#{pg.id}"][:minimum_cost] : number_to_currency(pp.minimum_cost, unit: "", delimiter: ""),
-      size: 8
-
-%td
-  = t("price_policies.amount")
-  %br
-  = text_field_tag "price_policy_#{pg.id}[cancellation_cost]",
-    params["price_policy_#{pg.id}"] ? params["price_policy_#{pg.id}"][:cancellation_cost] : number_to_currency(pp.cancellation_cost, unit: '', delimiter: ''),
-    size: 8
