--- app/views/price_policies/instrument/_internal_row.html.haml
@@ -1,20 +1,0 @@
-%td
-  = hidden_field_tag "price_policy_#{pg.id}[usage_rate]",
-    display_usage_rate(pg, pp),
-    class: "usage_cost"
-  = t("price_policies.adjustment")
-  %br
-  = text_field_tag "price_policy_#{pg.id}[usage_subsidy]",
-    display_usage_subsidy(pg, pp), size: 8, class: "usage_adjustment"
-
-%td
-  = t("price_policies.amount")
-  %br
-  %span.minimum_cost{ data: { usage_subsidy: pp.usage_subsidy } }
-  = hidden_price_policy_tag(pg.id, pp.minimum_cost, :minimum_cost, params["price_policy_#{pg.id}"])
-
-%td
-  = t("price_policies.amount")
-  %br
-  %span.cancellation_cost
-  = hidden_price_policy_tag(pg.id, pp.cancellation_cost, :cancellation_cost, params["price_policy_#{pg.id}"])
