--- app/views/price_policies/instrument/_master_internal_row.html.haml
@@ -1,21 +1,0 @@
-%td
-  = t("price_policies.rate")
-  %br
-  = text_field_tag "price_policy_#{pg.id}[usage_rate]",
-    display_usage_rate(pg, pp), size: 8, class: "master_usage_cost usage_rate"
-
-%td
-  = t("price_policies.amount")
-  %br
-  = text_field_tag "price_policy_#{pg.id}[minimum_cost]",
-    params["price_policy_#{pg.id}"] ? params["price_policy_#{pg.id}"][:minimum_cost] : number_to_currency(pp.minimum_cost, unit: "", delimiter: ""),
-    size: 8,
-    class: "master_minimum_cost"
-
-%td
-  = t("price_policies.amount")
-  %br
-  = text_field_tag "price_policy_#{pg.id}[cancellation_cost]",
-    params["price_policy_#{pg.id}"] ? params["price_policy_#{pg.id}"][:cancellation_cost] : number_to_currency(pp.cancellation_cost, unit: "", delimiter: ""),
-    size: 8,
-    class: "master_cancellation_cost"
