+++ app/views/price_policies/new.html.haml
@@ -11,12 +11,11 @@
 %h3= text("views.price_policies.index.add")
 %p= t("views.price_policies.edit.instructional_text")
 
-= modelless_form_for :url => price_policies_path do |f|
-  - @price_policies.each do |pp|
-    - @price_policy = pp
-    = error_messages_for :price_policy
+= modelless_form_for url: price_policies_path do |f|
+  - @price_policies.each do |price_policy|
+    = error_messages_for price_policy
 
-  = render "price_policies/#{@product.is_a?(Instrument) ? 'instrument/' : ''}price_policy_fields", f: f
+  = render "price_policy_fields", f: f
   %ul.inline
     %li= submit_tag text("views.price_policies.index.add"), class: "btn btn-primary"
     %li= link_to text("shared.cancel"), price_policies_path
